refactor: remove section-divider comments and migrate large inline test modules (#443)
Per the rule-05 style guide, section-marker comments (`// ---- title ----`)
are tolerated only when a function can't be split further. In test
modules they instead indicate the module has grown large enough to
benefit from a sibling `tests.rs` file or simple removal.

- `db/src/helpers.rs`: migrate the ~240-line inline `#[cfg(test)] mod tests`
  to a new sibling `db/src/helpers/tests.rs`, dropping the four
  `// ---------- ... ----------` section dividers.
- `db/src/library_layout.rs`: migrate the ~430-line inline test module to
  `db/src/library_layout/tests.rs`, dropping the four section dividers.
- `db/src/auth/password.rs`: drop the `// ---- username policy ---` divider.
- `db/src/auth.rs`: drop the three `// ---` re-export/section dividers.
- `db/src/author_photos_data.rs`: replace six horizontal-rule test-category
  dividers with blank lines.

Pure reorganization: no test bodies, assertions, or production logic were
changed. `cargo test -p omnibus-db` still reports 457 lib tests passing.

Closes #443
